Generate branch/tag script for ALL repositories
Previously, we only generated the branch/tag commands that Jenkins
itself needed to run (which is for most, but not all, repositories).

Now, we also generate the branch/tag commands for the repositories
that are handled manually (eg when Jenkins does not have access).
This means that the commands can be copy/pasted if required.