DASCTEST-373 Add org.mariadb.jdbc (MariaDB) to DLS p2 site
Note that this bundle needs to be available for both GDA 8 and GDA 9.
It was added to the "logging" p2 site, even though it has nothing to do with logging,
because GDA 8 and 9 use the same "logging" p2 site. The other p2 sites have diverged
and so are not suitable (in particular, the GDA 9 versions have had items removed
that are still used in GDA 8).