Update Squish runner to use Squish 6.3.0, and suppress the Dawn Welcome screen
In 6.3.0, the way Squish is packaged and installed was changed
(relative to the previous release we used), so script amended accordingly.

At some point, for reasons unknown, Dawn tests started crashing as soon
as the application started (JVM crash due to error in native code). The
dumps suggested JavaFX was involve, which Dawn uses for the Welcome screen,
so that was removed (for Squish test purposes), which stopped the crashes.