PEWMA: handle a rare intermittent bug resulting in false compile errors
The Buckminster materialize appears to work correctly, but fails to add
uk.ac.gda.api to the workspace (it's always that plugin). As a result,
a subsequent compile then gets 1000s of errors.

I have only ever seen this problem in Jenkins. This change attempts to
detect when this specific problem has occurred, and report that explicitly
in the Jenkins job description.